feat: separate component for Timeline
